Begin refactoring code into game
