+++ src/scripts/Game.js
@@ -1,10 +1,17 @@
-import Map from './Map'
-import Player from './Player';
-import Input from './Input'
-import Tree from './Tree';
+import Map from './scripts/Map';
+import Player from './scripts/Player';
+import Input from './scripts/Input';
+import Tree from './scripts/Tree';
+import Basket from './scripts/Basket';
+import Decoration from './scripts/Decoration';
 
 class Game {
-
+    constructor(canvWidth, canvHeight, mapWidth, mapHeight) {
+        this.canvWidth = canvWidth;
+        this.canvHeight = canvHeight;
+        this.mapWidth = mapWidth;
+        this.mapHeight = mapHeight;
+    }
 }
 
 export default Game;
